+++ src/main/java/com/Java24GroupProject/AirBnBPlatform/DTOs/ListingResponse.java
@@ -1,17 +1,15 @@
 package com.Java24GroupProject.AirBnBPlatform.DTOs;
 
-import com.Java24GroupProject.AirBnBPlatform.models.User;
 import com.Java24GroupProject.AirBnBPlatform.models.supportClasses.ListingUtilities;
 
-
-import javax.swing.text.Utilities;
+import java.math.BigDecimal;
 import java.util.Set;
 
 //The DTO for returning information at successful listing creation
 public class ListingResponse {
     private String title;
     private String description;
-    private Double price_per_night;
+    private BigDecimal price_per_night;
     private Integer capacity;
     private Set<ListingUtilities> utilities;
     private String host;
@@ -35,11 +33,11 @@ public class ListingResponse {
         this.description = description;
     }
     
-    public Double getPrice_per_night() {
+    public BigDecimal getPrice_per_night() {
         return price_per_night;
     }
     
-    public void setPrice_per_night(Double price_per_night) {
+    public void setPrice_per_night(BigDecimal price_per_night) {
         this.price_per_night = price_per_night;
     }
